Refactor JwsClaims to JwsClaimsNames object and JwsBaseClaims class
