+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/JwtBaseClaims.kt
@@ -1,0 +1,235 @@
+@file:UseSerializers(ByteArrayBase64Serializer::class)
+
+package at.asitplus.signum.indispensable.josef
+
+import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
+import at.asitplus.signum.indispensable.io.InstantLongSerializer
+import at.asitplus.signum.indispensable.josef.jwtpayload.ClientStatus
+import at.asitplus.signum.indispensable.josef.jwtpayload.ConfirmationClaim
+import at.asitplus.signum.indispensable.josef.jwtpayload.JwtClaimNames.IanaRegistered
+import at.asitplus.signum.indispensable.josef.jwtpayload.JwtClaimNames.UnregisteredClaims
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.UseSerializers
+import kotlinx.serialization.json.JsonObject
+import kotlin.time.Instant
+
+@Suppress("unused")
+@Deprecated("Renamed to better reflect content", ReplaceWith("jwtpayload"))
+typealias JsonWebToken = JwtBaseClaims
+
+/**
+ * Base-Claims defined in RFC7519
+ *
+ * See [RFC 7519](https://datatracker.ietf.org/doc/html/rfc7519)
+ */
+@Serializable
+data class JwtBaseClaims(
+
+    /**
+     * RFC 7519: The "iss" (issuer) claim identifies the principal that issued the
+     * JWT.  The processing of this claim is generally application specific.
+     * The "iss" value is a case-sensitive string containing a StringOrURI
+     * value.  Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.ISS)
+    val issuer: String? = null,
+
+    /**
+     * RFC 7519: The "sub" (subject) claim identifies the principal that is the
+     * subject of the JWT.  The claims in a JWT are normally statements
+     * about the subject.  The subject value MUST either be scoped to be
+     * locally unique in the context of the issuer or be globally unique.
+     * The processing of this claim is generally application specific.  The
+     * "sub" value is a case-sensitive string containing a StringOrURI
+     * value.  Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.SUB)
+    val subject: String? = null,
+
+    /**
+     * RFC 7519: The "aud" (audience) claim identifies the recipients that the JWT is
+     * intended for.  Each principal intended to process the JWT MUST
+     * identify itself with a value in the audience claim.  If the principal
+     * processing the claim does not identify itself with a value in the
+     * "aud" claim when this claim is present, then the JWT MUST be
+     * rejected.  In the general case, the "aud" value is an array of case-sensitive strings,
+     * each containing a StringOrURI value.  In the
+     * special case when the JWT has one audience, the "aud" value MAY be a
+     * single case-sensitive string containing a StringOrURI value.  The
+     * interpretation of audience values is generally application specific.
+     * Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.AUD)
+    val audience: String? = null,
+
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(IanaRegistered.ClaimNames.OpenIdConnectCore.NONCE)
+    val nonce: String? = null,
+
+    /**
+     * RFC 7519: The "nbf" (not before) claim identifies the time before which the JWT
+     * MUST NOT be accepted for processing.  The processing of the "nbf"
+     * claim requires that the current date/time MUST be after or equal to
+     * the not-before date/time listed in the "nbf" claim.  Implementers MAY
+     * provide for some small leeway, usually no more than a few minutes, to
+     * account for clock skew.  Its value MUST be a number containing a
+     * NumericDate value.  Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.NBF)
+    @Serializable(with = InstantLongSerializer::class)
+    val notBefore: Instant? = null,
+
+    /**
+     * RFC 7519: The "iat" (issued at) claim identifies the time at which the JWT was
+     * issued.  This claim can be used to determine the age of the JWT.  Its
+     * value MUST be a number containing a NumericDate value.  Use of this
+     * claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.IAT)
+    @Serializable(with = InstantLongSerializer::class)
+    val issuedAt: Instant? = null,
+
+    /**
+     * RFC 7519: The "exp" (expiration time) claim identifies the expiration time on
+     * or after which the JWT MUST NOT be accepted for processing.  The
+     * processing of the "exp" claim requires that the current date/time
+     * MUST be before the expiration date/time listed in the "exp" claim.
+     * Implementers MAY provide for some small leeway, usually no more than
+     * a few minutes, to account for clock skew.  Its value MUST be a number
+     * containing a NumericDate value.  Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.EXP)
+    @Serializable(with = InstantLongSerializer::class)
+    val expiration: Instant? = null,
+
+    /**
+     * RFC 7519: The "jti" (JWT ID) claim provides a unique identifier for the JWT.
+     * The identifier value MUST be assigned in a manner that ensures that
+     * there is a negligible probability that the same value will be
+     * accidentally assigned to a different data object; if the application
+     * uses multiple issuers, collisions MUST be prevented among values
+     * produced by different issuers as well.  The "jti" claim can be used
+     * to prevent the JWT from being replayed.  The "jti" value is a case-sensitive string.  Use of this claim is OPTIONAL.
+     */
+    @SerialName(IanaRegistered.ClaimNames.RFC7519.JTI)
+    val jwtId: String? = null,
+
+    /**
+     * OID4VP: This claim contains the confirmation method as defined in RFC7800. It MUST contain a JWK as defined in
+     * Section 3.2 of RFC7800. This claim determines the public key for which the corresponding private key the
+     * Verifier MUST proof possession of when presenting the Verifier Attestation JWT. This additional security measure
+     * allows the Verifier to obtain a Verifier Attestation JWT from a trusted issuer and use it for a long time
+     * independent of that issuer without the risk of an adversary impersonating the Verifier by replaying a captured
+     * attestation.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(IanaRegistered.ClaimNames.RFC7800.CNF)
+    val confirmationClaim: ConfirmationClaim? = null,
+
+    /**
+     * RFC 9449: The value of the HTTP method (Section 9.1 of [RFC9110](https://datatracker.ietf.org/doc/html/rfc9110))
+     * of the request to which the JWT is attached.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(IanaRegistered.ClaimNames.RFC9449.HTM)
+    val httpMethod: String? = null,
+
+    /**
+     * RFC 9449: The HTTP target URI (Section 7.1 of [RFC9110](https://datatracker.ietf.org/doc/html/rfc9110)) of the
+     * request to which the JWT is attached, without query and fragment parts.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(IanaRegistered.ClaimNames.RFC9449.HTU)
+    val httpTargetUrl: String? = null,
+
+    /**
+     * RFC 9449: Hash of the access token. The value MUST be the result of a base64url encoding (as defined in Section
+     * 2 of [RFC7515](https://datatracker.ietf.org/doc/html/rfc7515)) the SHA-256 hash of the ASCII encoding of the
+     * associated access token's value.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(IanaRegistered.ClaimNames.RFC9449.ATH)
+    val accessTokenHash: String? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. String containing a human-readable name of the Wallet.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(UnregisteredClaims.EudiTs3Claims.WALLET_NAME)
+    val walletName: String? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. String containing a URL to get further information about the Wallet and the Wallet Provider.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(UnregisteredClaims.EudiTs3Claims.WALLET_LINK)
+    val walletLink: String? = null,
+
+    /**
+     * EUDI TS3 WUA 1.5: version of the Wallet Solution.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(UnregisteredClaims.EudiTs3Claims.WALLET_VERSION)
+    val walletVersion: String? = null,
+
+    /**
+     * EUDI TS3 WUA 1.5: information about the certification achieved by the Wallet Solution.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(UnregisteredClaims.EudiTs3Claims.WALLET_SOLUTION_CERTIFICATION_INFORMATION)
+    val walletSolutionCertificationInformation: String? = null,
+
+    /**
+     * EUDI TS3 WUA 1.5: status list reference for the Wallet Instance and the time until which the Wallet Provider
+     * commits to maintaining the referenced status.
+     */
+    @Deprecated(
+        "Will be removed in the future. Please use specialized JWT-Claims class instead!",
+        level = DeprecationLevel.WARNING
+    )
+    @SerialName(UnregisteredClaims.EudiTs3Claims.CLIENT_STATUS)
+    val clientStatus: ClientStatus? = null,
+
+    /**
+     * OPTIONAL: Data class containing information for instance/unit attestation
+     * See: https://github.com/eu-digital-identity-wallet/eudi-doc-standards-and-technical-specifications/blob/main/docs/technical-specifications/ts3-wallet-unit-attestation.md
+     */
+    @Deprecated("TS3 WUA 1.5 removed eudi_wallet_info from Wallet Instance Attestations.")
+    @SerialName("eudi_wallet_info")
+    val eudiWalletInfo: EudiWalletInfo? = null,
+
+    /**
+     * OID4VCI: OPTIONAL. Status mechanism for the Wallet Attestation as defined in ietf-oauth-status-list.
+     */
+    @Deprecated("TS3 WUA 1.5 replaced top-level status with client_status.")
+    @SerialName("status")
+    val status: JsonObject? = null,
+)
