+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/jwtpayload/WalletAttestationPayload.kt
@@ -1,8 +1,11 @@
 package at.asitplus.signum.indispensable.josef.jwtpayload
 
 import at.asitplus.propigator.common.ObjectBackedValidated
-import at.asitplus.propigator.json.*
-import at.asitplus.signum.indispensable.josef.JwtClaims
+import at.asitplus.propigator.json.JsonBackingCodec
+import at.asitplus.propigator.json.JsonObjectBacked
+import at.asitplus.propigator.json.JsonObjectBackedSerializer
+import at.asitplus.propigator.json.jsonSlice
+import at.asitplus.signum.indispensable.josef.JwtBaseClaims
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
@@ -16,14 +19,14 @@ data class WalletAttestationPayload(
     private val json: Json = joseCompliantSerializer,
 ) : JsonObjectBacked(raw, JsonBackingCodec(json)), ObjectBackedValidated, JwtPayload {
 
-    val jwtClaims: JwtClaims by jsonSlice()
+    val jwtBaseClaims: JwtBaseClaims by jsonSlice()
     val walletAttestationClaims: WalletAttestationClaims by jsonSlice()
 
     override fun validate() {
-        jwtClaims
+        jwtBaseClaims
         walletAttestationClaims
-        jwtClaims.subject!!
-        jwtClaims.expiration!!
+        jwtBaseClaims.subject!!
+        jwtBaseClaims.expiration!!
     }
 
     object Serializer : KSerializer<WalletAttestationPayload> by JsonObjectBackedSerializer(::WalletAttestationPayload)
