+++ indispensable-josef/src/commonTest/kotlin/at/asitplus/signum/indispensable/josef/WalletAttestationJwtTest.kt
@@ -43,7 +43,7 @@ val WalletAttestationJwtTest by matrixSuite {
 
         val parsed: WalletAttestationPayload = joseCompliantSerializer.decodeFromString(input)
 
-        parsed.jwtClaims.issuer.shouldBeNull()
+        parsed.jwtBaseClaims.issuer.shouldBeNull()
         parsed.walletAttestationClaims.walletName shouldBe "Wallet Solution X by Wonderland State Department"
         parsed.walletAttestationClaims.walletVersion shouldBe "1.2.3"
         parsed.walletAttestationClaims.walletLink shouldBe "https://example.com/wallet/detail_info.html"
